Create edit.js for Media
