+++ assets/src/edit-story/elements/image/display.js
@@ -26,7 +26,7 @@ import { useRef } from 'react';
 import StoryPropTypes from '../../types';
 import { elementFillContent, getMediaSizePositionProps } from '../shared';
 import { useTransformHandler } from '../../components/transform';
-import { imageWithScale, getImageWithScaleCss } from './util';
+import { mediaWithScale, getMediaWithScaleCss } from '../media/util';
 
 const Element = styled.div`
   ${elementFillContent}
@@ -35,7 +35,7 @@ const Element = styled.div`
 
 const Img = styled.img`
   position: absolute;
-  ${imageWithScale}
+  ${mediaWithScale}
 `;
 
 function ImageDisplay({
@@ -68,7 +68,7 @@ function ImageDisplay({
           focalX,
           focalY
         );
-        target.style.cssText = getImageWithScaleCss(newImgProps);
+        target.style.cssText = getMediaWithScaleCss(newImgProps);
       }
     }
   });
