+++ assets/src/edit-story/elements/image/edit.js
@@ -15,158 +15,13 @@
  */
 
 /**
- * External dependencies
- */
-import styled from 'styled-components';
-import { useCallback, useState } from 'react';
-
-/**
  * Internal dependencies
  */
-import {
-  elementFillContent,
-  elementWithFlip,
-  CropBox,
-  getMediaSizePositionProps,
-  EditPanMovable,
-  ScalePanel,
-  MEDIA_MASK_OPACITY,
-} from '../shared';
-import { useStory } from '../../app';
 import StoryPropTypes from '../../types';
-import WithMask from '../../masks/display';
-import getTransformFlip from '../shared/getTransformFlip';
-import EditCropMovable from '../shared/editCropMovable';
-import { imageWithScale } from './util';
-
-const Element = styled.div`
-  ${elementFillContent}
-`;
-
-// Opacity of the mask is reduced depending on the opacity assigned to the image.
-const FadedImg = styled.img`
-  position: absolute;
-  opacity: ${({ opacity }) =>
-    opacity ? opacity * MEDIA_MASK_OPACITY : MEDIA_MASK_OPACITY};
-  pointer-events: none;
-  ${imageWithScale}
-  ${elementWithFlip}
-`;
-
-// Opacity is adjusted so that the double image opacity would equal
-// the opacity assigned to the image.
-const CropImg = styled.img`
-  position: absolute;
-  opacity: ${({ opacity }) =>
-    opacity ? 1 - (1 - opacity) / (1 - opacity * MEDIA_MASK_OPACITY) : null};
-  ${imageWithScale}
-  ${elementWithFlip}
-`;
+import MediaEdit from '../media/edit';
 
 function ImageEdit({ element, box }) {
-  const {
-    id,
-    resource,
-    opacity,
-    scale,
-    flip,
-    focalX,
-    focalY,
-    isFill,
-    isBackground,
-  } = element;
-  const { x, y, width, height, rotationAngle } = box;
-
-  const [fullImage, setFullImage] = useState(null);
-  const [croppedImage, setCroppedImage] = useState(null);
-  const [cropBox, setCropBox] = useState(null);
-
-  const {
-    actions: { updateElementById },
-  } = useStory();
-  const setProperties = useCallback(
-    (properties) => updateElementById({ elementId: id, properties }),
-    [id, updateElementById]
-  );
-
-  const imgProps = getMediaSizePositionProps(
-    resource,
-    width,
-    height,
-    scale,
-    flip?.horizontal ? 100 - focalX : focalX,
-    flip?.vertical ? 100 - focalY : focalY
-  );
-
-  imgProps.transformFlip = getTransformFlip(flip);
-
-  return (
-    <Element>
-      <FadedImg
-        ref={setFullImage}
-        draggable={false}
-        src={resource.src}
-        {...imgProps}
-        opacity={opacity / 100}
-      />
-      <CropBox ref={setCropBox}>
-        <WithMask element={element} fill={true} applyFlip={false}>
-          <CropImg
-            ref={setCroppedImage}
-            draggable={false}
-            src={resource.src}
-            {...imgProps}
-            opacity={opacity / 100}
-          />
-        </WithMask>
-      </CropBox>
-
-      {fullImage && croppedImage && (
-        <EditPanMovable
-          setProperties={setProperties}
-          fullMedia={fullImage}
-          croppedMedia={croppedImage}
-          flip={flip}
-          x={x}
-          y={y}
-          width={width}
-          height={height}
-          rotationAngle={rotationAngle}
-          offsetX={imgProps.offsetX}
-          offsetY={imgProps.offsetY}
-          mediaWidth={imgProps.width}
-          mediaHeight={imgProps.height}
-        />
-      )}
-
-      {!isFill && !isBackground && cropBox && croppedImage && (
-        <EditCropMovable
-          setProperties={setProperties}
-          cropBox={cropBox}
-          croppedImage={croppedImage}
-          flip={flip}
-          x={x}
-          y={y}
-          width={width}
-          height={height}
-          rotationAngle={rotationAngle}
-          offsetX={imgProps.offsetX}
-          offsetY={imgProps.offsetY}
-          imgWidth={imgProps.width}
-          imgHeight={imgProps.height}
-        />
-      )}
-
-      <ScalePanel
-        setProperties={setProperties}
-        x={x}
-        y={y}
-        width={width}
-        height={height}
-        scale={scale || 100}
-      />
-    </Element>
-  );
+  return <MediaEdit element={element} box={box} />;
 }
 
 ImageEdit.propTypes = {
